fix borders glitches in lighttable only
basically, I discovered that thumbnails height and width were computed and used internally as float in expose_filemanager() but passed to dt_image_expose() as int32. This caused rounding issues and misplacement of group borders, when redrawing of background over group lines.

This changes width and height to float, use fixed-sized borders and accurately move the thumbnails and borders where they should.

TODO: port the changes to zoomable table view and filmstrips.

lighttable : rework UI for preview & culling
infos part only

+++ src/views/view.h
@@ -189,8 +189,8 @@ typedef struct dt_view_image_expose_t
   dt_view_image_over_t *image_over;
   uint32_t imgid;
   cairo_t *cr;
-  int32_t width;
-  int32_t height;
+  float width;
+  float height;
   int32_t zoom;
   int32_t px;
   int32_t py;
